Add if statement while removing ipy_mongo profile. Needed to successfully build old and new DAS releases (where this file is removed)

+++ das.spec
@@ -14,7 +14,9 @@ Requires: py2-sphinx py2-pycurl rotatelogs
 %setup -D -T -b 1 -n DAS
 
 # remove ipython deps
+if [ ! -f src/python/DAS/tools/ipy_profile_mongo.py ]; then
 rm src/python/DAS/tools/ipy_profile_mongo.py
+fi
 
 %build
 cd ../WMCore
